Node.replaceChildAtIdx doesn't work with cloning
We need to be able to cache fragments and clone them, and therefore we
cannot cache the elements that contain references as direct children.
Instead we will have to create a td runtime method that takes an element
(the base fragment) and an array of indexes that will dictate where the
insertion should happen.

The array of indexes is working correctly now.