Allow promises values to be inserted much later
After the documentfragment is appended to the DOM it is basically discarded by the
browser, and the children it once contained are gone. That means that
trying to look up elements starting with the document fragment will not
work after the documentFragment is appended.

To solve this, we store a reference to the oldNode (the node to be
replaced) and use that to do the replacement.