Don't cache HTML, cache node refs instead
Because we are cloning nodes, we can't cache references to dom nodes.
Instead, we have to do a lookup to find the nodes at render time. This
commit enables us to cache the dom nodes that will be replaced by
dynamic values, and eliminates the need for the lookup.

This fixes a bug where node lookups are incorrect if previously inserted
dynamic values include multiple nodes. The hope is that the performance
lost from having to rebuild the HTML for each render will be offset (and
perhaps outweighed) by not having to do the index path lookups and by
not having to clone document fragments.