Add looping helpers and helperContext
Looping helpers require a globalish context (so they can know if they
are first, last, etc.) The helperContext is a place where helpers can
store information that can be used by other helpers (e.g. nested
helpers) or within the template (e.g. @let could be used to add
constants--they're not really variables--to Tornado, that can be used as
references in other places in a Tornado template). This is a rough
draft.