Making compiler extensions easier
Nothing is really different here, except that the compiler passes are
added with helpers. Each helper can multiple passes to each stage of
compilation.

Two new methods are added to the compiler: `useHelper` and `useHelpers`.
The first allows the compiler to use a given compiler helper, and does
the same for multiple helpers.