Turn sections params into helper context values
Because Tornado does not allow you to look up the context, there needs
to be a way to reference data outside of a sections's context. This can
now be done with section params.

The section params are added to the helper context, and can be
referenced by using the param name, prefixed with a `$`. For example:

```
{#data outerValue=myVal}
  {$outerValue}
{/data}
```

The helper context can now hold objects and arrays instead of just
scalar values, and lookups on the helper context uses the same function
as lookups on the normal context (which means functions are called and
Promises are handled the same way).