Allow alternative path separators in references
One common complaint about Dust is that dots in a reference are always
interpreted as steps in the context path. This change allows for dots in
the path to be interpreted as plain text strings instead of steps in the
context path.

Example:

```
{com.linkedin.dots.in.path|myVal sep="|"}
```

```
{
  "com.linkedin.dots.in.path": {
    "myVal": "DOTS"
  }
}
```

TODO: Make this work for helpers as well (currently only works for
references)