remove ctx state in favor of a setStatePass and a stateStack
the state of escapable raw and html attributes and tornado body are now
in a setState pass. The setState pass puts a state onto a node.

we introduce a stateStack which collects states from the nodes as we
build the instruction. the stateStack is an array of states we were in
as we walked the tree to build instructions.

remove dist add prepublish and npmignore