Text in textareas is plaintext
Fixes #141. Text in textareas is parsed as a plainsrting with the
exception of tornado elements, which are still parsed. There is still
the problem with an HTML element inside of a tornado body inside of a
text area. I think this solution may be the right path for that bug as
well.